Fully implements componentDidMount
This fully implements componentDidMount. In the process I refactored the
way everything within the window component is structured, splitting them
into mixins the way that skatejs does.

This was primarily done (in addition to make smaller, more
understandable pieces) so that we could conditionally add mixin behavior
*only* when needed. For now this is only done for the componentDidMount
functionality.

This means, though, that we won't send unnecessary messages back and
forth to tell the worker component that a mount occurred when the
component doesn't have a `componentDidMount` method. This saves on the
legwork that the mixin does as well.

In the future we might be able to do the same for some of the other
functionality we have (or add). Not sure if any of the event stuff could
be conditional or not.