Adds a garbage collector for handles
Each time a render occurs we queue a GC event, which will go through all
of the handles for an instance and mark them as not in use. This is the
sweep phase. The next time a collection occurs items marked as not in
use are removed from the instance map and overall handle store.

Fixes #30